workspace/fairsplit/src/backend/requirements.txt
# Runtime dependencies only — installed in Docker image
# Dev/CI tools (ruff, pytest, pip-audit) live in requirements-dev.txt

# Web framework + ASGI server
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Settings
pydantic-settings==2.6.1

# Auth
PyJWT==2.10.1

# Observability
prometheus-fastapi-instrumentator==7.0.0

# HTTP client (used in smoke_test.py and internal calls)
httpx==0.28.1
